documentation

+++ app/Http/Controllers/WishlistContentsController.php
@@ -2,6 +2,8 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Controllers\Api\Auth\Controller;
+use App\Models\Wishlist;
 use App\Models\WishlistContents;
 use App\Traits\ApiResponse;
 use Illuminate\Http\Request;
@@ -17,17 +19,6 @@ class WishlistContentsController extends Controller
     use ApiResponse;
     
     /**
-     * list of the user wishlist contents
-     *
-     * @return JsonResponse
-     */
-    public function index()
-    {
-        $wishlists_contents = WishlistContents::all();
-        return $this->successResponse($wishlists_contents);
-    }
-    
-    /**
      * Store (create) a product in an existing wishlist
      *
      * @param  Request $request
@@ -35,24 +26,19 @@ class WishlistContentsController extends Controller
      */
     public function store(Request $request)
     {
+        $wishlist = Wishlist::find($request->input('wishlist_id'));
+        
+        if (!$wishlist) {
+           return $this->errorResponse('Unknown wishlist');
+        }
+
+        $user_id = $request->input('user_id');
+        if ($wishlist->id != $user_id) {
+            return $this->errorResponse('Item is not public');
+        }
+
         $data = $request->input();
         $wishlistContents = WishlistContents::create($data);
-        return $this->successResponse($wishlistContents);
-    }
-    
-    /**
-     * Updates (patch) a product in an existing wishlist
-     *
-     * @param  Request $request
-     * @param  integer $id
-     * @return JsonResponse
-     */
-    public function update(Request $request, $id)
-    {
-        $wishlistContents = WishlistContents::find($id);
-        $data = $request->input();
-        $wishlistContents->update($data);
-
         return $this->successResponse($wishlistContents);
     }
     
@@ -62,9 +48,20 @@ class WishlistContentsController extends Controller
      * @param integer $id
      * @return JsonResponse
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
         $wishlistContents = WishlistContents::find($id);
+        $wishlist = Wishlist::find($wishlistContents->wishlist_id);
+
+        if (!$wishlist) {
+            return $this->errorResponse('Unknown wishlist');
+         }
+ 
+         $user_id = $request->input('user_id');
+         if ($wishlist->id != $user_id) {
+             return $this->errorResponse('Item is not public');
+         }
+         
         $wishlistContents->delete();
         return $this->successResponse(null);
     }
